requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.2.1